Move the yearly subscription tests to its own file
